+++ compass/core/logon.py
@@ -6,16 +6,16 @@ import certifi
 from lxml import html
 import requests
 
-from compass import schemas
-from compass.errors import CompassAuthenticationError
-from compass.errors import CompassError
-from compass.interface_base import InterfaceBase
-from compass.logging import logger
-from compass.settings import Settings
-from compass.utility import cast
-from compass.utility import compass_restify
-from compass.utility import PeriodicTimer
-from compass.utility import setup_tls_certs
+from compass.core import schemas
+from compass.core.errors import CompassAuthenticationError
+from compass.core.errors import CompassError
+from compass.core.interface_base import InterfaceBase
+from compass.core.logging import logger
+from compass.core.settings import Settings
+from compass.core.utility import cast
+from compass.core.utility import compass_restify
+from compass.core.utility import PeriodicTimer
+from compass.core.utility import setup_tls_certs
 
 TYPES_UNIT_LEVELS = Literal["Group", "District", "County", "Region", "Country", "Organisation"]
