+++ compass/core/schemas/member.py
